+++ src/tyr/serializers.cc
@@ -123,7 +123,8 @@ std::string serializeStatus(Api& request) {
   if (request.status().has_has_transit_tiles_case())
     status_doc.AddMember("has_transit_tiles",
                          rapidjson::Value().SetBool(request.status().has_transit_tiles()), alloc);
-  if (request.status().has_osm_changeset_case())
+  // a 0 changeset indicates there's none, so don't write in the output
+  if (request.status().has_osm_changeset_case() && request.status().osm_changeset())
     status_doc.AddMember("osm_changeset",
                          rapidjson::Value().SetUint64(request.status().osm_changeset()), alloc);